The debian live creation tools since an update beginning of december  reports for netboot creation:
"Bootloader in this image type not yet supported by live-build."
"This would produce a not bootable image, aborting (FIXME)."

In order to fix this, the netboot_fix.sh script which is temporarily
 downgrading to a november snapshot is created.
 It also reverts back to the current version
after creation of the netbootimage.

Additional script netboot_install.sh is an example script which makes additional settings for providing PXE boot.
Server ip 192.168.2.10 is assumed.